requirements.txt
Keras==3.4.1
Keras_Preprocessing==1.1.2
matplotlib==3.9.1
pandas==2.2.2
nltk==3.2.5
numpy==2.0.1
tensorflow==2.17.0
scikit_learn==1.5.1
pymongo==4.8.0
Flask==3.0.3
python-dotenv==0.0.5
wget==3.2